+++ client/package.json
@@ -28,11 +28,9 @@
   },
   "dependencies": {
     "craco": "^0.0.3",
-    "craco-alias": "^3.0.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.1",
     "socket.io-client": "^4.5.1",
     "styled-components": "^5.3.5",
     "styled-normalize": "^8.0.7"
@@ -49,6 +47,8 @@
     "@types/socket.io-client": "^3.0.0",
     "@types/styled-components": "^5.1.25",
     "typescript": "^4.7.2",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "craco-alias": "^3.0.1",
+    "react-scripts": "5.0.1"
   }
 }
