+++ client/src/components/common/Header/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { HeaderContainer } from "@components/common/Header/Header.style";
+import { HeaderContainer } from "./Header.style";
 const Header = ({ children }: { children: JSX.Element }) => {
   return <HeaderContainer>{children}</HeaderContainer>;
 };
