+++ client/src/components/common/index.tsx
@@ -1,0 +1,5 @@
+import DropBox from "./DropBox";
+import Header from "./Header";
+import Modal from "./Modal";
+
+export { DropBox, Header, Modal };
